+++ inc/models/Simple2DCell.h
@@ -60,6 +60,9 @@ class Simple2DCell : public Simple2DModel
         //!allow for cell division, according to a vector of model-dependent parameters
         virtual void cellDivision(const vector<int> &parameters,const vector<Dscalar> &dParams={});
 
+        //!allow for cell death, killing off the cell with the specified index
+        virtual void cellDeath(int cellIndex);
+
         //!Set cell positions according to a user-specified vector
         void setCellPositions(vector<Dscalar2> newCellPositions);
         //!Set vertex positions according to a user-specified vector
@@ -146,6 +149,11 @@ class Simple2DCell : public Simple2DModel
         GPUArray<int> vertexCellNeighbors;
         //! CELL neighbors of every cell
         GPUArray<int> cellNeighbors;
+        //!The number of vertices defining each cell
+        /*!
+        cellVertexNum[c] is an integer storing the number of vertices that make up the boundary of cell c.
+        */
+        GPUArray<int> cellVertexNum;
 
         //!an array containing net force on each vertex
         GPUArray<Dscalar2> vertexForces;
@@ -166,6 +174,23 @@ class Simple2DCell : public Simple2DModel
         //!The current energy of the system; only updated when an explicit energy calculation is called (i.e. not by default each timestep)
         Dscalar Energy;
         //!To write consistent files...the cell that started the simulation as index i has current index tagToIdx[i]
+        /*!
+         The Hilbert sorting stuff makes keeping track of particles, and re-indexing things when
+         particle number changes, a pain. Here's a description of the four relevant data structures.
+         tagToIdx[i] = a. At the beginning of a simulation, a particle had index "i", meaning its
+                        current state was found in position "i" of all various data vectors and arrays.
+                        That same particle's data is now in position "a" of those data structures.
+                        Short version: "Where do I look to find info for what I orinally called partice i?"
+        idxToTag[a] = i. That is, idxToTag just helps invert the tagToIdx list.
+                    idxToTag[tagToIdx[i]]=i
+        The above two structures (and the vertex versions of them) tell you how to go back and forth
+        between the current state of the system and the initial state of the system. What about going
+        back and forth between the current sorted state and the previous sorted state? The "itt" and
+        "tti" vectors give this information.
+        The itt and tti vectors are completely overwritten each time a spatial sorting is called.
+        By the way, I apologize if the nomenclature of "index" vs. "tag" is the opposite of what you,
+        the reader of these code comments, might expect.
+        */
         vector<int> tagToIdx;
         //!To write consistent files...the vertex that started the simulation as index i has current index tagToIdx[i]
         vector<int> tagToIdxVertex;
@@ -193,11 +218,6 @@ class Simple2DCell : public Simple2DModel
         GPUArray<Dscalar2> AreaPeri;//(current A,P) for each cell
         //!The area and perimeter preferences of each cell
         GPUArray<Dscalar2> AreaPeriPreferences;//(A0,P0) for each cell
-        //!The number of vertices defining each cell
-        /*!
-        cellVertexNum[c] is an integer storing the number of vertices that make up the boundary of cell c.
-        */
-        GPUArray<int> cellVertexNum;
         //!The number of CELL neighbors of each cell. For simple models this is the same as cellVertexNum, but does not have to be
         GPUArray<int> cellNeighborNum;
         //!A structure that indexes the vertices defining each cell