+++ src/models/voronoiModelBase.cpp
@@ -918,6 +918,39 @@ void voronoiModelBase::allDelSets()
         getDelSets(ii);
     };
 
+/*!
+A utility function for resizing data arrays... used by the cell division and cell death routines
+*/
+void voronoiModelBase::resizeAndReset()
+    {
+    //Simple resizing operations
+    displacements.resize(Ncells);
+    cellForces.resize(Ncells);
+    external_forces.resize(Ncells);
+    exclusions.resize(Ncells);
+    NeighIdxs.resize(6*(Ncells+10));
+    circumcenters.resize(2*(Ncells+10));
+    repair.resize(Ncells);
+
+    celllist.setNp(Ncells);
+    resetDelLocPoints();
+    cellNeighborNum.resize(Ncells);
+    //brute force fix of triangulation
+    globalTriangulationCGAL();
+    resetLists();
+    allDelSets();
+    };
+
+/*!
+Trigger a cell death event. In the Voronoi model this simply removes the targeted cell and instantaneously computes the new tesselation. Very violent, if the cell isn't already small.
+*/
+void voronoiModelBase::cellDeath(int cellIndex)
+    {
+    //first, call the parent class routines.
+    //This call already changes Ncells
+    Simple2DActiveCell::cellDeath(cellIndex);
+    resizeAndReset();
+    };
 
 /*!
 Trigger a cell division event, which involves some laborious re-indexing of various data structures.
@@ -1018,21 +1051,5 @@ void voronoiModelBase::cellDivision(const vector<int> &parameters, const vector<
     cp.data[cellIdx] = newCellPos1;
     cp.data[Ncells-1] = newCellPos2;
     }
-
-    //Simple resizing operations
-    displacements.resize(Ncells);
-    cellForces.resize(Ncells);
-    external_forces.resize(Ncells);
-    exclusions.resize(Ncells);
-    NeighIdxs.resize(6*(Ncells+10));
-    circumcenters.resize(2*(Ncells+10));
-    repair.resize(Ncells);
-
-    celllist.setNp(Ncells);
-    resetDelLocPoints();
-    cellNeighborNum.resize(Ncells);
-    //brute force fix of triangulation
-    globalTriangulationCGAL();
-    resetLists();
-    allDelSets();
+    resizeAndReset();
     };